fix: text file preview crash, add MD tables and code toggle
- Fix SyntaxHighlighter crash (duplicate .r/.R key in case-insensitive dict)
- Add markdown table rendering with header styling
- Add </> toggle for rendered/code view on MD and HTML files
- Add HTML preview type with WebView2 rendering
- Fix UTF-16 encoded files falsely detected as binary
- Exclude emoji popup from screen capture during recording
- Bump version to 6.5.0

+++ Llamashot/Core/AviWriter.cs
@@ -86,8 +86,8 @@ public class AviWriter : IDisposable
         _writer.Write(0); // suggested buffer size
         _writer.Write(-1); // quality
         _writer.Write(0); // sample size
-        _writer.Write((short)0); _writer.Write((short)0); // frame rect
-        _writer.Write((short)_width); _writer.Write((short)_height);
+        _writer.Write(0); _writer.Write(0); // frame rect (left, top)
+        _writer.Write(_width); _writer.Write(_height); // frame rect (right, bottom)
 
         // strf - stream format (BITMAPINFOHEADER)
         WriteChunkId("strf");
